modified:   Assignments/Programming Assignment 1 (due 09182012)/.idea/Programming Assignment 1 (due 09182012).iml 	new file:   Assignments/Programming Assignment 1 (due 09182012)/Marr_Hildreth_algorithm.py 	new file:   Assignments/Programming Assignment 1 (due 09182012)/PA1.py 	new file:   Assignments/Programming Assignment 1 (due 09182012)/__pycache__/PA1.cpython-39.pyc 	modified:   Assignments/Programming Assignment 1 (due 09182012)/main.py
